refactor: Add explicit CourseLecture entity
Change ManyToMany implementation to use two ManyToOne relationships in
CourseLecture class. Course.type will be thus be implemented in service
